Add files via upload
Certificate of Participation for actively participating in the 2nd National Level Hackathon – HackArena, organized by the IEEE Student Branch at Guru Nanak Institutions, Hyderabad. Represented CMR College of Engineering and Technology as a member of Team Sustainovaters, where we developed and presented an innovative solution under the theme "Clean and Green."
The event provided a great platform to collaborate with peers, sharpen problem-solving skills, and gain valuable hands-on experience in real-time project development.